driver/util: rename util pkg to driverutil
