+++ drivers/driverutil/util.go
@@ -1,4 +1,4 @@
-package util
+package driverutil
 
 import (
 	"fmt"
