+++ drivers/driverutil/util_test.go
@@ -1,4 +1,4 @@
-package util
+package driverutil
 
 import (
 	"testing"
